Fixed bug for uint8_t use for leds which can be higher than 128.

+++ LedStrip.cpp
@@ -60,7 +60,7 @@ struct FastLedCRGB* LedStrip::GetLed(uint8_t ledIndex)
 
 void LedStrip::SetAllLeds(LedColor::EColor color, uint8_t step)
 {
-	for (uint8_t ledIndex = 0; ledIndex < _nrOfLeds; ledIndex++)
+	for (uint16_t ledIndex = 0; ledIndex < _nrOfLeds; ledIndex++)
 	{
 		FastLedCRGB& led = _leds[ledIndex];
 		LedColor::SetRgb(&led.red, &led.green, &led.blue, color, step);
@@ -70,7 +70,7 @@ void LedStrip::SetAllLeds(LedColor::EColor color, uint8_t step)
 
 void LedStrip::SetAllLeds(uint32_t color)
 {
-	for (uint8_t ledIndex = 0; ledIndex < _nrOfLeds; ledIndex++)
+	for (uint16_t ledIndex = 0; ledIndex < _nrOfLeds; ledIndex++)
 	{
 		FastLedCRGB& led = _leds[ledIndex];
 		LedColor::SetRgb(&led.red, &led.green, &led.blue, color);
@@ -80,7 +80,7 @@ void LedStrip::SetAllLeds(uint32_t color)
 
 void LedStrip::SetAllLeds(uint8_t red, uint8_t green, uint8_t blue)
 {
-	for (uint8_t ledIndex = 0; ledIndex < _nrOfLeds; ledIndex++)
+	for (uint16_t ledIndex = 0; ledIndex < _nrOfLeds; ledIndex++)
 	{
 		FastLedCRGB& led = _leds[ledIndex];
 		LedColor::SetRgb(&led.red, &led.green, &led.blue, red, green, blue);
@@ -109,7 +109,7 @@ void LedStrip::SetLed(struct FastLedCRGB* led, uint8_t red, uint8_t green, uint8
 uint16_t LedStrip::CalculateCurrentInMilliAmp()
 {
 	uint32_t brightness = 0;
-	for (uint8_t ledIndex = 0; ledIndex < _nrOfLeds; ledIndex++)
+	for (uint16_t ledIndex = 0; ledIndex < _nrOfLeds; ledIndex++)
 	{
 		FastLedCRGB& led = _leds[ledIndex];
 		brightness += led.red + led.green + led.blue;
